Change the byte of the MemoryAttributes attribute to the size type
There is a patch [1] about changing the byte type of the
MemoryAttributes attribute to size_t, so we need to update the
relevant code in bmcweb synchronously.

[1] https://gerrit.openbmc.org/c/openbmc/phosphor-dbus-interfaces/+/63799

+++ redfish-core/lib/memory.hpp
@@ -414,7 +414,7 @@ inline void
     const std::string* ecc = nullptr;
     const std::string* formFactor = nullptr;
     const std::vector<uint16_t>* allowedSpeedsMT = nullptr;
-    const uint8_t* memoryAttributes = nullptr;
+    const size_t* memoryAttributes = nullptr;
     const uint16_t* memoryConfiguredSpeedInMhz = nullptr;
     const std::string* memoryType = nullptr;
     const std::string* channel = nullptr;
@@ -531,8 +531,7 @@ inline void
 
     if (memoryAttributes != nullptr)
     {
-        asyncResp->res.jsonValue[jsonPtr]["RankCount"] =
-            static_cast<uint64_t>(*memoryAttributes);
+        asyncResp->res.jsonValue[jsonPtr]["RankCount"] = *memoryAttributes;
     }
 
     if (memoryConfiguredSpeedInMhz != nullptr)
